requirements.txt
Flask==2.0.2
flask-oidc==1.4.0
gunicorn==20.1.0
itsdangerous==2.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
python-dateutil==2.8.2
python-jose==3.3.0
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1